+++ packages/cli/src/handlers/compile.ts
@@ -24,12 +24,7 @@ import {
 import { validateDbtModel } from '../dbt/validation';
 import GlobalState from '../globalState';
 import * as styles from '../styles';
-import {
-    dbtCompileNecessaryModels,
-    DbtCompileOptions,
-    dbtList,
-    dbtParse,
-} from './dbt/compile';
+import { dbtCompile, DbtCompileOptions, dbtList } from './dbt/compile';
 import { getDbtVersion, isSupportedDbtVersion } from './dbt/getDbtVersion';
 
 export type CompileHandlerOptions = DbtCompileOptions & {
@@ -81,6 +76,13 @@ export const compile = async (options: CompileHandlerOptions) => {
 
     // Skipping assumes manifest.json already exists.
     let compiledModelIds: string[] | undefined;
+    if (options.useDbtList) {
+        compiledModelIds = await dbtList(options);
+    } else if (!options.skipDbtCompile) {
+        await dbtCompile(options);
+    } else {
+        GlobalState.debug('> Skipping dbt compile');
+    }
 
     const absoluteProjectPath = path.resolve(options.projectDir);
     const absoluteProfilesPath = path.resolve(options.profilesDir);
@@ -89,27 +91,6 @@ export const compile = async (options: CompileHandlerOptions) => {
     GlobalState.debug(`> Compiling with profiles dir ${absoluteProfilesPath}`);
 
     const context = await getDbtContext({ projectDir: absoluteProjectPath });
-
-    if (options.useDbtList) {
-        compiledModelIds = await dbtList(options);
-    } else if (!options.skipDbtCompile) {
-        // Generate manifest file without compiling any models
-        await dbtParse(dbtVersion, {
-            profilesDir: options.profilesDir,
-            projectDir: options.projectDir,
-            profile: options.profile,
-            target: options.target,
-        });
-
-        // Compile all models and their joins
-        await dbtCompileNecessaryModels(
-            { targetDir: context.targetDir },
-            options,
-        );
-    } else {
-        GlobalState.debug('> Skipping dbt compile');
-    }
-
     const profileName = options.profile || context.profileName;
     const { target } = await loadDbtTarget({
         profilesDir: absoluteProfilesPath,
@@ -127,10 +108,8 @@ export const compile = async (options: CompileHandlerOptions) => {
             ? options.startOfWeek
             : undefined,
     });
-
     const manifest = await loadManifest({ targetDir: context.targetDir });
-    const manifestModels = getModelsFromManifest(manifest);
-    const compiledModels = manifestModels.filter((model) => {
+    const models = getModelsFromManifest(manifest).filter((model) => {
         if (compiledModelIds) {
             return compiledModelIds.includes(model.unique_id);
         }
@@ -139,8 +118,9 @@ export const compile = async (options: CompileHandlerOptions) => {
     });
 
     const adapterType = manifest.metadata.adapter_type;
+
     const { valid: validModels, invalid: failedExplores } =
-        await validateDbtModel(adapterType, compiledModels);
+        await validateDbtModel(adapterType, models);
 
     if (failedExplores.length > 0) {
         const errors = failedExplores.map((failedExplore) =>