+++ packages/e2e/cypress/e2e/app/dashboard.cy.ts
@@ -168,13 +168,21 @@ describe('Dashboard', () => {
         // Check tile targets are correct and all charts have that filter applied
         cy.contains('Payment method is credit_card').click();
         cy.findAllByRole('tab').eq(1).click();
-        cy.get('.mantine-Checkbox-body').should('have.length', 4); // 3 checkboxes for the 3 charts + `select all` checkbox
-        cy.get('.mantine-Checkbox-body').each(($el) => {
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        ).should('have.length', 4); // 3 checkboxes for the 3 charts + `select all` checkbox
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        ).each(($el) => {
             cy.wrap($el).find('input[checked]').should('have.length', 1);
         });
 
         // Remove filter from first chart - saved chart
-        cy.get('.mantine-Checkbox-body').eq(1).click();
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        )
+            .eq(1)
+            .click();
         cy.contains('button', 'Apply').click({ force: true });
 
         // Saved chart should have no filter applied
@@ -206,16 +214,26 @@ describe('Dashboard', () => {
         // Open filter popover  and check that all charts have the filter applied except for the new one (which is referencing another explore)
         cy.contains('Payment method is credit_card').click();
         cy.findAllByRole('tab').eq(1).click();
-        cy.get('.mantine-Checkbox-body').should('have.length', 5); // 4 checkboxes for the 4 charts + `select all` checkbox
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        ).should('have.length', 5); // 4 checkboxes for the 4 charts + `select all` checkbox
 
         // Enable filter for the new chart
-        cy.get('.mantine-Checkbox-body').eq(4).click();
-        cy.get('.mantine-Checkbox-body')
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        )
+            .eq(4)
+            .click();
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        )
             .eq(4)
             .within(() => {
                 cy.get('input').should('be.checked');
             });
-        cy.get('.mantine-Checkbox-body')
+        cy.get(
+            '[data-testid="DashboardFilterConfiguration/ChartTiles"] .mantine-Checkbox-body',
+        )
             .eq(4)
             .parent()
             .parent()